@@ -4,7 +4,7 @@
   "description": "Example LOB application using Signum Framework and React",
   "main": "index.js",
   "scripts": {
-    "watch": "yarn run clean && webpack --env.NODE_ENV=development --config webpack.config.js --watch --progress",
+    "watch": "yarn run clean && webpack --env.NODE_ENV=development --config webpack.config.js --watch",
     "build": "yarn run clean && webpack --env.NODE_ENV=production --config webpack.config.js",
     "clean": "rimraf ./wwwroot/dist"
   },
@@ -25,52 +25,52 @@
   },
   "homepage": "https://github.com/signumsoftware/southwind",
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "1.2.18",
-    "@fortawesome/free-regular-svg-icons": "5.8.2",
-    "@fortawesome/free-solid-svg-icons": "5.8.2",
+    "@fortawesome/fontawesome-svg-core": "1.2.22",
+    "@fortawesome/free-regular-svg-icons": "5.10.2",
+    "@fortawesome/free-solid-svg-icons": "5.10.2",
     "@fortawesome/react-fontawesome": "0.1.4",
-    "@types/prop-types": "file:../Framework/Signum.React/node_modules/@types/prop-types",
-    "@types/react": "file:../Framework/Signum.React/node_modules/@types/react",
+    "@types/prop-types": "15.7.2",
+    "@types/react": "16.9.2",
     "abortcontroller-polyfill": "1.3.0",
     "bootstrap": "4.3.1",
     "bpmn-js": "2.5.2",
     "diagram-js-minimap": "1.3.0",
-    "codemirror": "5.47.0",
-    "d3": "5.9.2",
-    "d3-scale-chromatic": "1.3.3",
-    "es6-promise": "4.2.6",
+    "codemirror": "5.48.4",
+    "d3": "5.11.0",
+    "d3-scale-chromatic": "1.5.0",
+    "es6-promise": "4.2.8",
     "es6-object-assign": "1.1.0",
-    "history": "4.7.2",
-    "jquery": "3.3.1",
+    "history": "4.9.0",
+    "jquery": "3.4.1",
     "lodash-compat": "3.10.2",
     "moment": "2.24.0",
     "numbro": "2.1.2",
-    "popper.js": "1.14.7",
+    "popper.js": "1.15.0",
     "prop-types": "15.7.2",
-    "query-string": "6.5.0",
-    "react": "16.8.6",
-    "react-popper": "0.10.4",
-    "react-dom": "16.8.6",
-    "react-overlays": "0.8.3",
-    "react-router": "4.3.1",
-    "react-router-dom": "4.3.1",
+    "query-string": "6.8.3",
+    "react-bootstrap": "1.0.0-beta.10",
+    "react-router": "5.0.1",
+    "react-router-dom": "5.0.1",
+    "react": "16.9.0",
+    "react-dom": "16.9.0",
     "react-rte": "0.16.1",
     "react-widgets": "4.4.11",
+    "uncontrollable": "7.0.1",
     "whatwg-fetch": "3.0.0"
   },
   "devDependencies": {
     "assets-webpack-plugin": "4.0.0",
-    "css-loader": "2.1.1",
+    "css-loader": "3.2.0",
     "file-loader": "3.0.1",
     "raw-loader": "1.0.0",
     "rimraf": "2.6.3",
     "style-loader": "0.23.1",
-    "ts-loader": "6.0.2",
-    "typescript": "3.6.2",
-    "url-loader": "1.1.2",
-    "webpack": "4.32.2",
-    "webpack-cli": "3.3.2",
-    "webpack-notifier": "1.7.0"
+    "ts-loader": "6.1.2",
+    "typescript": "3.6.3",
+    "url-loader": "2.1.0",
+    "webpack": "4.40.2",
+    "webpack-cli": "3.3.9",
+    "webpack-notifier": "1.8.0"
   },
   "-vs-binding": {
     "ProjectOpened": [
