@@ -1,19 +1,16 @@
-﻿import * as React from 'react'
+import * as React from 'react'
 import * as Navigator from "@framework/Navigator"
 import * as DashboardClient from "@extensions/Dashboard/DashboardClient"
 import * as AuthClient from '@extensions/Authorization/AuthClient'
+import { useAPI } from '../../Framework/Signum.React/Scripts/Hooks';
 
-export default class Home extends React.Component<{}, { loaded: boolean }> {
+export default function Home() {
 
-  constructor(props: {}) {
-    super(props);
-    this.state = { loaded: false };
-  }
+  var [loaded, setLoaded] = React.useState(false);
+  React.useEffect(() => {
 
-  componentWillMount() {
     if (!AuthClient.currentUser()) {
       Navigator.history.push("~/publicCatalog");
-      return;
     }//PublicCatalog
 
     if (AuthClient.currentUser()) {
@@ -22,36 +19,38 @@ export default class Home extends React.Component<{}, { loaded: boolean }> {
           if (h)
             Navigator.history.push(`~/dashboard/${h.id}`);
           else
-            this.setState({ loaded: true });
+            setLoaded(true);
         });
     }
     else //Dashboard
-      this.setState({ loaded: true });
-  }
+      setLoaded(true);
 
-  render() {
-    if (!this.state.loaded)
-      return null;
+  }, []);
 
-    return (
-      <div>
+
+
+
+  if (!loaded)
+    return null;
+
+  return (
+    <div>
+      <br />
+      <div className="jumbotron">
+        <h1 className="display-4">Welcome to Signum React</h1>
         <br />
-        <div className="jumbotron">
-          <h1 className="display-4">Welcome to Signum React</h1>
-          <br />
-          <p className="lead">Southwind is a demo application from <a href="http://www.signumsoftware.com" title="Signum Software">Signum Software</a> based on Northwind database from Microsoft:</p>
-          <p>
-            To learn more about Signum Framework visit <a href="http://www.signumframework.com" title="Signum Framework">http://www.signumframework.com</a>.
+        <p className="lead">Southwind is a demo application from <a href="http://www.signumsoftware.com" title="Signum Software">Signum Software</a> based on Northwind database from Microsoft:</p>
+        <p>
+          To learn more about Signum Framework visit <a href="http://www.signumframework.com" title="Signum Framework">http://www.signumframework.com</a>.
           </p>
-          <p> To be effective in Signum React you will also need to know:</p>
-          <ul>
-            <li><a href="http://www.typescriptlang.org/" title="Typescript">Typescript</a></li>
-            <li><a href="https://facebook.github.io/react">React</a></li>
-            <li><a href="http://getbootstrap.com/" title="Bootstrap">Bootstrap</a></li>
-            <li><a href="http://webpack.com/" title="Webpack">Webpack</a></li>
-          </ul>
-        </div>
+        <p> To be effective in Signum React you will also need to know:</p>
+        <ul>
+          <li><a href="http://www.typescriptlang.org/" title="Typescript">Typescript</a></li>
+          <li><a href="https://facebook.github.io/react">React</a></li>
+          <li><a href="http://getbootstrap.com/" title="Bootstrap">Bootstrap</a></li>
+          <li><a href="http://webpack.com/" title="Webpack">Webpack</a></li>
+        </ul>
       </div>
-    );
-  }
+    </div>
+  );
 }
