@@ -1,14 +1,14 @@
 import * as React from 'react'
 import { ApplicationConfigurationEntity } from '../Southwind.Entities'
 import { ValueLine, EntityLine, EntityCombo, EntityList, EntityDetail, EntityStrip, EntityRepeater, TypeContext, RenderEntity } from '@framework/Lines'
-import { UncontrolledTabs, Tab } from '@framework/Components/Tabs';
+import { Tabs, Tab } from 'react-bootstrap';
 
 export default function ApplicationConfiguration(p : { ctx: TypeContext<ApplicationConfigurationEntity> }){
   const ctx = p.ctx;
   return (
     <div>
       <ValueLine ctx={ctx.subCtx(a => a.environment)} />
-      <UncontrolledTabs>
+      <Tabs id="appTabs">
         <Tab eventKey="tab" title={ctx.niceName(a => a.email)}>
           <RenderEntity ctx={ctx.subCtx(a => a.email)} />
           <EntityLine ctx={ctx.subCtx(a => a.smtpConfiguration)} />
@@ -19,7 +19,7 @@ export default function ApplicationConfiguration(p : { ctx: TypeContext<Applicat
         <Tab eventKey="auth" title={ctx.niceName(a => a.authTokens)}>
           <RenderEntity ctx={ctx.subCtx(a => a.authTokens)} />
         </Tab>
-      </UncontrolledTabs>
+      </Tabs>
     </div>
   );
 }
