@@ -1,5 +1,5 @@
 import * as React from 'react'
-import * as numbro from 'numbro'
+import numbro from 'numbro'
 import * as moment from 'moment'
 import { Dic } from '@framework/Globals'
 import * as Navigator from '@framework/Navigator'
