@@ -27,10 +27,10 @@ module.exports = env => {
       splitChunks: {
         cacheGroups: {
           vendor: {
-            test: /[\\/]node_modules[\\/](react|react-dom|react-router|react-widgets|react-popper|react-overlays|query-string|moment|numbro|d3)[\\/]/,
+            test: /[\\/]node_modules[\\/](react|react-dom|react-router|react-widgets|react-bootstrap|query-string|moment|numbro|d3)[\\/]/,
             name: 'vendor',
             chunks: 'all',
-          }
+          },
         }
       },
     },
